Decode System + Body IDs from search string, allow -b to replace BodyID
If a search string with a BodyID already encoded in it is passed in as
input, decode the SystemID and BodyID and print them separately - useful
to recover the BodyID from galaxy map search strings documented on forum
posts, e.g. Halleys Comet:

    $ ./edgalmap.py Wregoe AC-D d12-22020096
    Wregoe AC-D d12-0, Body 21
    Copied to clipboard: "Wregoe AC-D d12-22020096"

Also, allow the -b parameter to replace any such Body ID with anthoer.
Brings the functionality of the name to search string mode closer to
that of the address to search string mode.

    $ ./edgalmap.py Slegi CC-B d14-25165828 -b 23
    WARNING: Address already included BodyID 24, replacing with -b 23
    Slegi CC-B d14-4, Body 23
    Copied to clipboard: "Slegi CC-B d14-24117252"