Cope with sector lookup errors determining system ID from name
Sectors like Col 69 are based on an XYZ coord + radius, rather than the
Boxel. These aren't in PGSectorNames.json, so trying to look them up to
calculate the System + Body IDs fails. Make this error non-fatal so the
script can still output the extended galaxy map search address as we
used to - in particular, makes the Col 69 example in the README.md work
again.

Fixes regression introduced in commit:

    commit 7e2aaf9e7839713e180c83e4524ae5a6efb0e925
    Author: Ian Munsie <darkstarsword@gmail.com>
    Date:   Tue Sep 24 04:02:19 2024 +1000

	Print out system + body address when passed a system name